+++ packages/index.ts
@@ -1,3 +1,5 @@
-export { Badge } from "./badge";
-export { BadgeButton } from "./badge";
-export { Column, Table, Cell, TextCell, HeaderCell } from "./table";
+import { injectGlobalCss } from "./shared/styles/global";
+export { Badge, BadgeButton } from "./badge";
+export { Column, Table, Cell, TextCell, HeaderCell, NumberCell } from "./table";
+
+injectGlobalCss();