+++ packages/table/components/NumberCell.tsx
@@ -5,4 +5,5 @@ export default styled(Cell)`
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
+  text-align: right;
 `;